testsuite: move libflux tests to subdirectory
