+++ src/common/libflux/module.c
@@ -257,9 +257,6 @@ char *flux_modfind (const char *searchpath, const char *modname)
     return modpath;
 }
 
-/* It is not convenient to test these directly here.
- */
-#ifndef TEST_MAIN
 int flux_insmod_request_decode (zmsg_t *zmsg, char **path,
                                 int *argc, char ***argv)
 {
@@ -380,169 +377,6 @@ done:
     Jput (in);
     return rc;
 }
-#endif /* !TEST_MAIN */
-
-
-#ifdef TEST_MAIN
-
-#include "src/common/libtap/tap.h"
-
-void test_helpers (void)
-{
-    char *name, *path;
-
-    ok ((strcmpend ("foo.so", ".so") == 0),
-        "strcmpend matches .so");
-    ok ((strcmpend ("", ".so") != 0),
-        "strcmpend doesn't match empty string");
-
-    name = mod_service ("kvs");
-    like (name, "^cmb$",
-        "mod_service of kvs is cmb");
-    if (name)
-        free (name);
-
-    name = mod_service ("sched.backfill");
-    like (name, "^sched$",
-        "mod_service of sched.backfill is sched");
-    if (name)
-        free (name);
-
-    name = mod_service ("sched.backfill.priority");
-    like (name, "^sched.backfill$",
-        "mod_service of sched.backfill.priority is sched.backfill");
-    if (name)
-        free (name);
-
-    path = xasprintf ("%s/kvs/.libs/kvs.so", MODULE_PATH);
-    ok (access (path, F_OK) == 0,
-        "built kvs module is located");
-    name = flux_modname (path);
-    ok ((name != NULL),
-        "flux_modname on kvs should find a name");
-    skip (name == NULL, 1,
-        "skip next test because kvs.so name is NULL");
-    like (name, "^kvs$",
-        "flux_modname says kvs module is named kvs");
-    end_skip;
-    if (name)
-        free (name);
-    ok (flux_modname_cmp (name, "kvs"),
-        "flux_modname_cmp also says kvs module is named kvs");
-    free (path);
-
-    ok (!modfind ("nowhere", "foo"),
-        "modfind fails with nonexistent directory");
-    ok (!modfind (".", "foo"),
-        "modfind fails in current directory");
-    ok (!modfind (MODULE_PATH, "foo"),
-        "modfind fails to find unknown module in moduledir");
-
-    path = xasprintf ("%s/kvs/.libs", MODULE_PATH);
-    name = modfind (path, "kvs");
-    ok ((name != NULL),
-        "modfind finds kvs in flat directory");
-    if (name)
-        free (name);
-    free (path);
-
-    name = modfind (MODULE_PATH, "kvs");
-    ok ((name != NULL),
-        "modfind finds kvs in deep moduledir");
-    if (name)
-        free (name);
-
-    name = flux_modfind (MODULE_PATH, "kvs");
-    ok ((name != NULL),
-        "flux_modfind also finds kvs in moduledir");
-    if (name)
-        free (name);
-
-    path = xasprintf ("foo:bar:xyz:%s:zzz", MODULE_PATH);
-    name = flux_modfind (path, "kvs");
-    ok ((name != NULL),
-        "flux_modfind also finds kvs in search path");
-    if (name)
-        free (name);
-    free (path);
-}
-
-void test_lsmod_codec (void)
-{
-    JSON o;
-    int len;
-    int idle, size;
-    const char *name, *digest;
-
-    o = flux_lsmod_json_create ();
-    ok (o != NULL,
-        "flux_lsmod_json_create works");
-    ok (flux_lsmod_json_append (o, "foo", 42, "aa", 3) == 0,
-        "first flux_lsmod_json_append works");
-    ok (flux_lsmod_json_append (o, "bar", 43, "bb", 2) == 0,
-        "second flux_lsmod_json_append works");
-    ok (flux_lsmod_json_decode (o, &len) == 0 && len == 2,
-        "flux_lsmod_json_decode works");
-    ok (flux_lsmod_json_decode_nth (o, 0, &name, &size, &digest, &idle) == 0
-        && name && size == 42 && digest && idle == 3
-        && !strcmp (name, "foo") && !strcmp (digest, "aa"),
-        "flux_lsmod_json_decode_nth(0) works");
-    ok (flux_lsmod_json_decode_nth (o, 1, &name, &size, &digest, &idle) == 0
-        && name && size == 43 && digest && idle == 2
-        && !strcmp (name, "bar") && !strcmp (digest, "bb"),
-        "flux_lsmod_json_decode_nth(1) works");
-
-    Jput (o);
-}
-
-void test_rmmod_codec (void)
-{
-    JSON o;
-    char *s = NULL;
-
-    o = flux_rmmod_json_encode ("xyz");
-    ok (o != NULL,
-        "flux_rmmod_json_encode works");
-    ok (flux_rmmod_json_decode (o, &s) == 0 && s != NULL && !strcmp (s, "xyz"),
-        "flux_rmmod_json_decode works");
-    free (s);
-    Jput (o);
-}
-
-void test_insmod_codec (void)
-{
-    int ac, argc = 2;
-    char *argv[] = { "foo", "bar" };
-    char **av;
-    JSON o;
-    char *s;
-
-    o = flux_insmod_json_encode ("/foo/bar", argc, argv);
-    ok (o != NULL,
-        "flux_insmod_json_encode works");
-    ok (flux_insmod_json_decode (o, &s, &ac, &av) == 0
-        && s != NULL && !strcmp (s, "/foo/bar")
-        && ac == 2 && !strcmp (av[0], "foo") && !strcmp (av[1], "bar"),
-        "flux_insmod_json_decode works");
-    argv_destroy (ac, av);
-    free (s);
-    Jput (o);
-}
-
-int main (int argc, char *argv[])
-{
-    plan (26);
-
-    test_helpers (); // 16
-    test_lsmod_codec (); // 6
-    test_rmmod_codec (); // 2
-    test_insmod_codec (); // 2
-
-    done_testing ();
-}
-
-
-#endif
 
 /*
  * vi:tabstop=4 shiftwidth=4 expandtab
